+++ src/Container/Exception/ContainerException.php
@@ -2,17 +2,17 @@
 
 namespace Rougin\Slytherin\Container\Exception;
 
-use Psr\Container\ContainerExceptionInterface;
+use Rougin\Slytherin\Container\ContainerException as BaseException;
 
 /**
  * Container Exception
  *
  * A specified exception in handling errors in containers.
- * NOTE: To be removed in v1.0.0. Use NotFoundException instead.
+ * NOTE: To be removed in v1.0.0. Use Container\ContainerException instead.
  *
  * @package Slytherin
  * @author  Rougin Gutib <rougingutib@gmail.com>
  */
-class ContainerException extends \Exception implements ContainerExceptionInterface
+class ContainerException extends BaseException
 {
 }